Added a Signal to ChipScope and Regenerated RAM
I think all my code for the data_send moudle is working now.  However, I
am not seeing the peaks in the signal when I look at the waveform.  I
kind of think that something is wrong with the RAM and that as I write
to it, it puts the information I write on the output bus as soon as it
is written.  I am not sure, but I think that might be it.  I regenerated
the core just to be certian.  I also added a few signals into chipscope
so that I can see, among other things, the data comign in and out of the
ram.  If this data is the same, then I know something is up.  In order
to add the signals to ChipScope, I needed more room, so I cut the
ethernet RAM in half and the chipScope core as well.  I think it will be
better and easier to manage this way.