Added Manual Trigger and Manual Mode
I added the code for the manual trigger as well as code for manual mode.
During normal opperation in threshold mode, you can force a trigger by
going to address 600 and typing in 1.  This will force the device to
trigger.  If there is a threshold trigger that is reading out before the
forced trigger, then it will take precidence over the forced trigger and
the forced trigger will be read out after the threshold one.  In manual
mode, the threshold functionality is disabled and only the manual
trigger works.