Gotcha! twice?  I think I actually found the error
Ah, so the last error was where I was doing if(reading <= '1').  That
shouldn't have had the assignment operator.  I fixed it and thought I
was done with the problem.  Unfortunantly, I just found another instance
of the same problem in the if statement that actually assigns one to
reading.  I think by fixing this one, I will fix the poblem.  I also did
a search for "if(*<=*)" and didn't come up with anything, so I think I
got all the bad "if" statements.  This commit COMPILED.