Added code for actual delay with positive delay
When I programmed in the positive delay, I simply added the number of
positive delay cycles onto the starting address to make it start later.
However, this was a bad idea in that any positive delay more than one or
two makes it start reading before the buffer has circled around.  I
added a counter and actual delay to delay for the number of positive
delay cycles so that the buffer will actually circle around before it
starts reading.  This commit COMPILED