Forgot to reset "busy" flag
When I created the busy flag for data_send, I forgot to set it and reset
it.  Last commit, I noticed that I never set it, so I put in the code to
set it.  Now everything is malfunctioning and I am pretty sure that it
is because I forgot to reset it.  I just added the reset code.  This
commit COMPILED.