Added Clock_Enable
I figured out that not all of the data coming from the adc_fifo is
actually valid.  Since the ADC is going slower than the analysis
firmware, it can't keep up and skipps a clock once in a while.  This
Clock_Enable signal is really the data_valid signal and will freeze
everything when the data coming from the ADC is not valid.