Wrote code for IP address changes
I wrote the code that allows us to set the IP addresses using resistors
on the Captan board.  I should also not that I fixed the bug where the
Captan wasn't triggering on an external trigger that happened during a
clk_enable low.  I think I have this fixed.  However, there is the issue
that if one trigger comes directly after the other, then there is the
chance that the firmware will miss it as it has been disabled the whole
time and the two triggers will just act as one.  I suppose I could add
this into the missed triggers count and output it with the rest of the
header, but it should be fine.